Remove prim::Constant op
Constants are interpreter primitives so the op was not actually used.
This cleans up some of the logic around it.

This also fixes constant prop such that failures to look up an op
do not silently stop constant propagation. Instead, only errors
inside the op implementation itself will do this.

ghstack-source-id: 7cf702d485730e95a24391490801fc98b3e9ae97
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32804